src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Homework Problem Solver Backend"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=5.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
]
